+++ src/main/java/kr/hhplus/be/server/coupon/domain/dto/reponse/CouponIssueResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.coupon.dto.reponse;
+package kr.hhplus.be.server.coupon.domain.dto.reponse;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
