+++ src/main/java/kr/hhplus/be/server/coupon/infrastructure/CouponPolicyJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.coupon.infrastructure.repository;
+package kr.hhplus.be.server.coupon.infrastructure;
 
 import kr.hhplus.be.server.coupon.domain.CouponPolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
