+++ src/main/java/kr/hhplus/be/server/coupon/infrastructure/CouponRepositoryImpl.java
@@ -2,7 +2,6 @@ package kr.hhplus.be.server.coupon.infrastructure;
 
 import kr.hhplus.be.server.coupon.domain.Coupon;
 import kr.hhplus.be.server.coupon.domain.CouponRepository;
-import kr.hhplus.be.server.coupon.infrastructure.repository.CouponJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
