+++ src/main/java/kr/hhplus/be/server/payment/domain/PaymentRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.payment.domain;
+
+public interface PaymentRepository {
+    Payment save(Payment payment);
+}
