+++ src/main/java/kr/hhplus/be/server/payment/infrastructure/PaymentJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.payment.infrastructure.repository;
+package kr.hhplus.be.server.payment.infrastructure;
 
 import kr.hhplus.be.server.payment.domain.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
