+++ src/main/java/kr/hhplus/be/server/product/application/ProductService.java
@@ -4,9 +4,9 @@ import kr.hhplus.be.server.common.exception.EntityNotFoundException;
 import kr.hhplus.be.server.common.exception.StockRecoveryException;
 import kr.hhplus.be.server.order.domain.OrderItem;
 import kr.hhplus.be.server.product.domain.Product;
-import kr.hhplus.be.server.product.domain.repository.ProductRepository;
-import kr.hhplus.be.server.product.dto.response.ProductDetailResponse;
-import kr.hhplus.be.server.product.dto.response.ProductResponse;
+import kr.hhplus.be.server.product.domain.ProductRepository;
+import kr.hhplus.be.server.product.domain.dto.response.ProductDetailResponse;
+import kr.hhplus.be.server.product.domain.dto.response.ProductResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.stereotype.Service;
