+++ src/main/java/kr/hhplus/be/server/product/infrastructure/ProductJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.product.infrastructure.repository;
+package kr.hhplus.be.server.product.infrastructure;
 
 import kr.hhplus.be.server.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
