+++ src/main/java/kr/hhplus/be/server/user/domain/PointHistoryRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.user.domain;
+
+public interface PointHistoryRepository{
+    void save(PointHistory history);
+}
