--- src/main/java/kr/hhplus/be/server/user/domain/repository/PointHistoryRepository.java
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.user.domain.repository;
-
-import kr.hhplus.be.server.user.domain.PointHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PointHistoryRepository{
-    void save(PointHistory history);
-}
