+++ src/main/java/kr/hhplus/be/server/user/infrastructure/UserJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.user.infrastructure.repository;
+package kr.hhplus.be.server.user.infrastructure;
 
 import kr.hhplus.be.server.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
