+++ src/main/java/kr/hhplus/be/server/user/infrastructure/UserRepositoryImpl.java
@@ -1,8 +1,7 @@
 package kr.hhplus.be.server.user.infrastructure;
 
 import kr.hhplus.be.server.user.domain.User;
-import kr.hhplus.be.server.user.domain.repository.UserRepository;
-import kr.hhplus.be.server.user.infrastructure.repository.UserJpaRepository;
+import kr.hhplus.be.server.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import java.util.Optional;
