+++ src/test/java/kr/hhplus/be/server/coupon/application/CouponServiceTest.java
@@ -2,8 +2,9 @@ package kr.hhplus.be.server.coupon.application;
 
 import kr.hhplus.be.server.coupon.domain.Coupon;
 import kr.hhplus.be.server.coupon.domain.CouponPolicy;
-import kr.hhplus.be.server.coupon.domain.repository.CouponPolicyRepository;
-import kr.hhplus.be.server.coupon.domain.repository.CouponRepository;
+
+import kr.hhplus.be.server.coupon.domain.CouponPolicyRepository;
+import kr.hhplus.be.server.coupon.domain.CouponRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
