+++ src/test/java/kr/hhplus/be/server/product/application/ProductServiceTest.java
@@ -3,9 +3,9 @@ package kr.hhplus.be.server.product.application;
 import kr.hhplus.be.server.common.exception.InsufficientStockException;
 import kr.hhplus.be.server.order.domain.OrderItem;
 import kr.hhplus.be.server.product.domain.Product;
-import kr.hhplus.be.server.product.domain.repository.ProductRepository;
-import kr.hhplus.be.server.product.dto.response.ProductDetailResponse;
-import kr.hhplus.be.server.product.dto.response.ProductResponse;
+import kr.hhplus.be.server.product.domain.ProductRepository;
+import kr.hhplus.be.server.product.domain.dto.response.ProductDetailResponse;
+import kr.hhplus.be.server.product.domain.dto.response.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -61,12 +61,12 @@ public class ProductServiceTest {
 
         // 주문 항목 설정
         orderItem1 = new OrderItem();
-        orderItem1.setProductId(1L);
-        orderItem1.setOrderQty(10);
+        orderItem1.setOrderedProductId(1L);
+        orderItem1.setOrderItemQty(10);
 
         orderItem2 = new OrderItem();
-        orderItem2.setProductId(2L);
-        orderItem2.setOrderQty(20);
+        orderItem2.setOrderedProductId(2L);
+        orderItem2.setOrderItemQty(20);
 
         orderItems = Arrays.asList(orderItem1, orderItem2);
